backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Bruno Faigle-Cedzich"]
readme = "README.md"

[tool.poetry.scripts]
generate = "app.engine.generate:generate_datasource"

[tool.poetry.dependencies]
python = "^3.12"
llama-index-llms-ollama = "^0.3.2"
jupyter = "^1.1.1"
fastapi = "^0.115.0"
cachetools = "^5.5.0"
llama-index-embeddings-ollama = "^0.3.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
